Forcefully fix set_local_themedef override issue
The issue is that the data of an imported module is shared across all
other modules that use it, meaning that the alt_dir and related
variables in frontend will be updated across everywhere. I forcefully
created a workaround by including all string files in all
set_local_themedef uses.